handles duplicate requests
NOTES:
an `ArrayList` is used to check if the user has asked to grade the
same course, it informs the user and prompts again for a different
course tag. Once the grading is done for all the courses it quits.

Note that we have yet to add the grading logic.